Move  ByteShuffler to it's own module.
